optimize util files
